requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
pillow